+++ lib/src/core/request_context.dart
@@ -59,69 +59,14 @@ abstract class RequestContext<RawRequest> {
   /// The original HTTP verb sent to the server.
   String get originalMethod;
 
-  StateError _unparsed(String type, String caps) => new StateError(
-      'Cannot get the $type of an unparsed request. Use lazy${caps}() instead.');
-
-  /// All post data submitted to the server.
-  ///
-  /// If you are lazy-parsing request bodies, but have not manually [parse]d this one,
-  /// then an error will be thrown.
-  ///
-  /// **If you are writing a plug-in, use [parseBody] instead.**
-  Map get body {
-    if (_body == null)
-      throw _unparsed('body', 'Body');
-    else
-      return _body.body;
-  }
-
   /// The content type of an incoming request.
   MediaType get contentType;
-
-  /// Any and all files sent to the server with this request.
-  ///
-  /// If you are lazy-parsing request bodies, but have not manually [parse]d this one,
-  /// then an error will be thrown.
-  ///
-  /// **If you are writing a plug-in, use [parseUploadedFiles] instead.**
-  List<FileUploadInfo> get files {
-    if (_body == null)
-      throw _unparsed('query', 'Files');
-    else
-      return _body.files;
-  }
-
-  /// The original body bytes sent with this request. May be empty.
-  ///
-  /// If you are lazy-parsing request bodies, but have not manually [parse]d this one,
-  /// then an error will be thrown.
-  ///
-  /// **If you are writing a plug-in, use [parseRawRequestBuffer] instead.**
-  List<int> get originalBuffer {
-    if (_body == null)
-      throw _unparsed('original buffer', 'OriginalBuffer');
-    else
-      return _body.originalBuffer ?? [];
-  }
 
   /// The URL parameters extracted from the request URI.
   Map params = {};
 
   /// The requested path.
   String get path;
-
-  /// The parsed request query string.
-  ///
-  /// If you are lazy-parsing request bodies, but have not manually [parse]d this one,
-  /// then [uri].query will be returned.
-  ///
-  /// **If you are writing a plug-in, consider using [lazyQuery] instead.**
-  Map get query {
-    if (_body == null)
-      return _provisionalQuery ??= new Map.from(uri.queryParameters);
-    else
-      return _body.query;
-  }
 
   /// The remote address requesting this resource.
   InternetAddress get remoteAddress;
@@ -223,7 +168,7 @@ abstract class RequestContext<RawRequest> {
   /// Retrieves the request body if it has already been parsed, or lazy-parses it before returning the query.
   ///
   /// If [forceParse] is not `true`, then [uri].query will be returned, and no parsing will be performed.
-  Future<Map<String, dynamic>> lazyQuery({bool forceParse: false}) {
+  Future<Map<String, dynamic>> parseQuery({bool forceParse: false}) {
     if (_body == null && forceParse != true)
       return new Future.value(uri.queryParameters);
     else