+++ CHANGELOG.md
@@ -27,3 +27,4 @@ with `Angel.eagerParseRequestBodies`.
 * `Angel.storeOriginalBuffer` -> `Angel.storeRawRequestBuffers`.
 * The methods `lazyBody`, `lazyFiles`, and `lazyOriginalBuffer` on `ResponseContext` were all
 replaced with `parseBody`, `parseUploadedFiles`, and `parseRawRequestBuffer`, respectively.
+* Removed the synchronous equivalents of the above methods (`body`, `files`, and `originalBuffer`).